further remove or move methods in commonutils.
