+++ core/src/main/java/tachyon/web/WebInterfaceBrowseLogsServlet.java
@@ -86,7 +86,7 @@ public class WebInterfaceBrowseLogsServlet extends HttpServlet {
           // stream couldn't read anything, skip went to EOF?
           fileData = "Unable to read file";
         } else {
-          fileData = CommonUtils.convertByteArrayToStringWithoutEscape(data, 0, read);
+          fileData = Utils.convertByteArrayToStringWithoutEscape(data, 0, read);
         }
       }
     } finally {
